move closestScreenshot logic into renderer
